+++ CustomSabers/Menu/Components/SaberListTableData.cs
@@ -70,8 +70,8 @@ internal class SaberListTableData : MonoBehaviour, TableView.IDataSource
     {
         if (scrollbar == null) return;
         var buttonBase = scrollbar.transform.Find("UpButton").GetComponent<Button>();
-        topButton = buttonBase.CreateButton(7f, new(0.5f, 1.0f), new(0.5f, 1.0f), new(2.5f, 2.5f), 180f, CSLResources.ExtremeArrowIcon, ScrollToTop, buttonBase.transform.parent);
-        bottomButton = buttonBase.CreateButton(7f, new(0.5f, 0f), new(0.5f, 0f), new(2.5f, 2.5f), 0f, CSLResources.ExtremeArrowIcon, ScrollToBottom, buttonBase.transform.parent);
+        topButton = CreateButton(buttonBase, 7f, new(0.5f, 1.0f), new(0.5f, 1.0f), new(2.5f, 2.5f), 180f, PluginResources.ExtremeArrowIcon, ScrollToTop, buttonBase.transform.parent);
+        bottomButton = CreateButton(buttonBase, 7f, new(0.5f, 0f), new(0.5f, 0f), new(2.5f, 2.5f), 0f, PluginResources.ExtremeArrowIcon, ScrollToBottom, buttonBase.transform.parent);
         tableView.scrollView.scrollPositionChangedEvent += ScrollPositionChanged;
     }
 
@@ -153,5 +153,45 @@ internal class SaberListTableData : MonoBehaviour, TableView.IDataSource
     {
         Normal = 0,
         Simple
+    }
+    
+    private static Button CreateButton(
+        Button original,
+        float size,
+        Vector2 anchorMin,
+        Vector2 anchorMax,
+        Vector2 iconSize,
+        float rotation,
+        Sprite? icon = null,
+        Action? onClick = null,
+        Transform? parent = null)
+    {
+        var button = Instantiate(original);
+        if (parent != null) button.transform.SetParent(parent, false);
+        if (onClick != null) button.onClick.AddListener(onClick.Invoke);
+        button.interactable = true;
+
+        var buttonTransform = (RectTransform)button.transform;
+        buttonTransform.anchorMin = anchorMin;
+        buttonTransform.anchorMax = anchorMax;
+        buttonTransform.sizeDelta = new(size, size);
+        buttonTransform.offsetMin = new(size / -2f, size / -2f);
+        buttonTransform.offsetMax = new(size / 2f, size / 2f);
+
+        var iconImageView = button.GetComponentInChildren<ImageView>();
+        if (icon != null) iconImageView.sprite = icon;
+
+        var iconTransform = (RectTransform)iconImageView.transform;
+        iconTransform.anchorMin = Vector2.zero;
+        iconTransform.anchorMax = Vector2.one;
+        iconTransform.anchorMin = new(0.5f, 0.5f);
+        iconTransform.anchorMax = new(0.5f, 0.5f);
+        iconTransform.offsetMin = Vector2.zero;
+        iconTransform.offsetMax = Vector2.zero;
+        iconTransform.sizeDelta = iconSize;
+        iconTransform.localEulerAngles = new(0, 0, rotation);
+        
+        button.gameObject.SetActive(true);
+        return button;
     }
 }
